+++ src/v1/components/hooks/use-books.ts
@@ -1,11 +1,18 @@
 import ApiV1 from '@/v1/api'
+import useBookselfContext from '@/v1/libs/context'
 import queryKeys from '@/v1/libs/query-keys'
 import {useQuery} from '@tanstack/react-query'
 
 function useBooks() {
+    const {
+        state: {
+            filter
+        },
+    } = useBookselfContext()
+
     return useQuery({
-        queryKey: queryKeys.books(),
-        queryFn: () => ApiV1.Book.query(),
+        queryKey: queryKeys.books(filter),
+        queryFn: () => ApiV1.Book.query(filter),
     })
 }