+++ src/v1/libs/query-keys.ts
@@ -1,6 +1,8 @@
+import {FilterType} from '@/v1/libs/types'
+
 const queryKeys = {
     default: ['bookself', 'v1'] as const,
-    books: () => [...queryKeys.default, 'books'],
+    books: (filter: FilterType) => [...queryKeys.default, 'books', filter],
 } as const
 
 export default queryKeys