+++ src/v1/libs/reducer.ts
@@ -1,14 +1,16 @@
-import {BookType, StateType} from '@/v1/libs/types'
+import {BookType, FilterType, StateType} from '@/v1/libs/types'
 import {getDefaultState} from '@/v1/libs/utils'
 import {useReducer} from 'react'
 
 enum ActionType {
     SET_BOOK = 'SET_BOOK',
     SET_BOOKS = 'SET_BOOKS',
+    SET_FILTER = 'SET_FILTER',
 }
 
 type Action =
     | { type: ActionType.SET_BOOK, payload: BookType | undefined }
+    | { type: ActionType.SET_FILTER, payload: FilterType }
 
 function reducer(prevState: StateType, action: Action): StateType {
     const {type, payload} = action
@@ -18,6 +20,12 @@ function reducer(prevState: StateType, action: Action): StateType {
             return {
                 ...prevState,
                 book: payload,
+            }
+
+        case ActionType.SET_FILTER:
+            return {
+                ...prevState,
+                filter: payload,
             }
 
         default: